Moved form fields around, deleted old registration templates
