+++ contentcuration/contentcuration/frontend/shared/app.js
@@ -10,6 +10,7 @@ import 'shared/i18n/setup';
 import 'vuetify/dist/vuetify.min.css';
 import 'shared/styles/main.less';
 import Base from 'shared/Base.vue';
+import ActionLink from 'shared/views/ActionLink';
 
 import { initializeDB } from 'shared/data';
 
@@ -25,6 +26,9 @@ Vue.use(Vuetify, {
   theme: theme(),
   icons: icons(),
 });
+
+// Register global components
+Vue.component('ActionLink', ActionLink);
 
 export let rootVue;
