--- contentcuration/contentcuration/templates/registration/activate.html
@@ -1,16 +1,0 @@
-{% extends "base.html" %}
-{% load staticfiles %}
-{% load i18n %}
-{% block nav %}{% endblock nav %}
-{% block content %}
-<div class="message-page">
-	<div class="pull-left message-img-wrapper">
-		<img class="message-logo" src="{% static 'img/kolibri_login.png' %}"/>
-	</div>
-	<div class="message-wrapper">
-		<br/>
-		<br/>
-		<h2>{% trans "Failed to activate account. Please try again." %}</h2>
-	</div>
-</div>
-{% endblock %}
