--- contentcuration/contentcuration/templates/registration/password_change_done.html
@@ -1,16 +1,0 @@
-{% extends "base.html" %}
-{% load staticfiles %}
-{% load i18n %}
-{% block nav %}{% endblock nav %}
-
-{% block content %}
-<div class="message-page">
-	<div class="pull-left message-img-wrapper">
-		<img class="message-logo" src="{% static 'img/kolibri_login.png' %}"/>
-	</div>
-	<div class="message-wrapper">
-		<h2>{% trans "Password changed successfully." %}</h2>
-		<a href="/" class="btn message-btn">{% trans "Continue to home..." %}</a>
-	</div>
-</div>
-{% endblock %}
