--- contentcuration/contentcuration/templates/registration/password_change_form.html
@@ -1,30 +1,0 @@
-<!DOCTYPE HTML>
-{% extends "base.html" %}
-{% load staticfiles %}
-{% load i18n %}
-{% block nav %}{% endblock nav %}
-
-{% block content %}
-<div class="message-page">
-	<div class="message-img-wrapper">
-		<img class="message-logo" src="{% static 'img/kolibri_login.png' %}"/>
-	</div>
-	<div class="message-wrapper">
-		<form method="post" action="/accounts/password/change/">
-			{% csrf_token %}
-			<h2>{% trans "Password Change" %}</h2>
-			{% if form.errors %}
-				<ul>
-					{% blocktrans %}{{ form.email.errors }}{% endblocktrans %}
-				</ul>
-			{% endif %}
-			<p>
-				{% csrf_token %}
-				{% trans "New password" %}{{form.new_password1}}
-				{% trans "New password confirmation" %}{{form.new_password2}}
-				<input class="btn message-btn" type="submit" value="{% trans 'Submit' %}" />
-			</p>
-		</form>
-	</div>
-</div>
-{% endblock %}
