--- contentcuration/contentcuration/templates/registration/password_reset_complete.html
@@ -1,17 +1,0 @@
-{% extends "base.html" %}
-{% load staticfiles %}
-{% load i18n %}
-{% block nav %}{% endblock nav %}
-
-{% block content %}
-<div class="message-page">
-	<div class="pull-left message-img-wrapper">
-		<img class="message-logo" src="{% static 'img/kolibri_login.png' %}"/>
-	</div>
-	<div class="message-wrapper">
-		<h2>{% trans "Password reset successfully" %}</h2>
-		<p>{% trans "Your password has been set. You may go ahead and log in now." %}</p>
-		<p><a href="{% url 'login' %}" class="btn message-btn">{% trans "Log in" %}</a></p>
-	</div>
-</div>
-{% endblock %}
