--- contentcuration/contentcuration/templates/registration/password_reset_done.html
@@ -1,17 +1,0 @@
-{% extends "base.html" %}
-{% load staticfiles %}
-{% load i18n %}
-{% block nav %}{% endblock nav %}
-
-{% block content %}
-<div class="message-page">
-	<div class="pull-left message-img-wrapper">
-		<img class="message-logo" src="{% static 'img/kolibri_login.png' %}"/>
-	</div>
-	<div class="message-wrapper">
-		<h2>{% trans "Password reset email sent." %}</h2>
-		<p>{% trans "We've emailed you instructions for setting your password, if an account exists with the email you entered. You should receive them shortly." %}</p>
-		<p>{% trans "If you don't receive an email, please make sure you've entered the address you registered with, and check your spam folder." %}</p>
-	</div>
-</div>
-{% endblock %}
