+++ contentcuration/contentcuration/templates/registration/password_reset_email.txt
@@ -6,7 +6,7 @@
 
 {% trans "Please go to the following page and choose a new password:" %}
 {% block reset_link %}
-{{ domain }}{% url 'auth_password_reset_confirm' uidb64=uid token=token %}
+{{ domain }}{% url 'accounts' %}#/reset-password?uidb64={{uid}}&token={{token}}
 {% endblock %}
 
 {% trans "Your username, in case you've forgotten:" %} {{ user.email }}
