--- contentcuration/contentcuration/templates/registration/registration_closed.html
@@ -1,14 +1,0 @@
-{% extends "base.html" %}
-{% load staticfiles %}
-{% load i18n %}
-{% block nav %}{% endblock nav %}
-{% block content %}
-<div class="message-page">
-	<div class="pull-left message-img-wrapper">
-		<img class="message-logo" src="{% static 'img/kolibri_login.png' %}"/>
-	</div>
-	<div class="message-wrapper">
-		<h2>{% trans "Registration is temporarily closed. Please check back later." %}</h2>
-	</div>
-</div>
-{% endblock %}
